Auth enable Endpoint test
