+++ backend/src/test/java/com/osoc6/OSOC6/adminTest/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * This module contains all tests that test with access level admin.
+ * It includes some utility classes so writing tests is a breeze.
+ *
+ * @author jitsedesmet
+ */
+
+package com.osoc6.OSOC6.adminTest;
