+++ backend/src/test/java/com/osoc6/OSOC6/coachTest/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This module contains all tests that test with access level coach.
+ *
+ * @author jitsedesmet
+ */
+
+package com.osoc6.OSOC6.coachTest;
