+++ backend/src/test/java/com/osoc6/OSOC6/package-info.java
@@ -1,5 +1,5 @@
 /**
- * This module contains all tests.
+ * This package contains all tests including some utility classes so writing tests is a breeze.
  *
  * @author ruvmello, jitsedesmet
  */
